+++ src/chrono/fea/ChNodeFEAxyzDDD.h
@@ -59,7 +59,7 @@ class ChApi ChNodeFEAxyzDDD : public ChNodeFEAxyzDD {
     ChVariables& Variables_DDD() { return *variables_DDD; }
 
     /// Reset to no speed and acceleration.
-    virtual void SetNoSpeedNoAcceleration() override;
+    virtual void SetZeroVelocityZeroAcceleration() override;
 
     /// Get mass of the node (for DDD variables).
     ChVectorDynamic<>& GetMassDiagonalDDD() { return variables_DDD->GetMassDiagonal(); }